some style and refuct mask functions and validation functions
